+++ app/src/_domain/auth/_auth.personal.page.tsx
@@ -1,0 +1,7 @@
+import Template from "./_auth.personal.template";
+
+function AuthPersonal(): JSX.Element {
+  return <Template />;
+}
+
+export default AuthPersonal;
