--- app/src/_domain/auth/_auth_personal.page.tsx
@@ -1,7 +1,0 @@
-import Template from "./_auth_personal.template";
-
-function AuthPersonal(): JSX.Element {
-  return <Template />;
-}
-
-export default AuthPersonal;
