+++ app/src/_domain/auth/_personal.login.page.tsx
@@ -1,0 +1,7 @@
+import Template from "./_personal.login.template";
+
+function PersonalLogin(): JSX.Element {
+  return <Template />;
+}
+
+export default PersonalLogin;
