+++ app/src/router/authRoutes.tsx
@@ -1,6 +1,7 @@
 import { RouteObject } from "react-router-dom";
 import AuthType from "../_domain/auth/_auth.type.page";
-import AuthPersonal from "../_domain/auth/_auth_personal.page";
+import AuthPersonal from "../_domain/auth/_auth.personal.page";
+import PersonalLogin from "../_domain/auth/_personal.login.page";
 
 const authRoutes: RouteObject[] = [
   {
@@ -10,6 +11,10 @@ const authRoutes: RouteObject[] = [
   {
     path: "/authPersonal",
     element: <AuthPersonal />,
+  },
+  {
+    path: "/login",
+    element: <PersonalLogin />,
   },
 ];
 export default authRoutes;
